chore: move sidetree-v2 to sidetree
