+++ packages/element-lib/index.js
@@ -4,7 +4,7 @@ const crypto = require('./src/crypto');
 const schema = require('./src/schema');
 const adapters = require('./src/adapters');
 const Sidetree = require('./src/sidetree');
-const SidetreeV2 = require('./src/sidetree-v2');
+const SidetreeV2 = require('./src/sidetree');
 const op = require('./src/sidetree/op');
 const MnemonicKeySystem = require('./src/crypto/MnemonicKeySystem');
