--- packages/element-lib/src/sidetree-v2/protocol/index.js
@@ -1,9 +1,0 @@
-const resolve = require('./resolve');
-const sync = require('./sync');
-const batchWrite = require('./batchWrite');
-
-module.exports = {
-  resolve,
-  sync,
-  batchWrite,
-};
