+++ packages/element-lib/src/sidetree/__tests__/sidetreeCreatePayload.json
