+++ packages/element-lib/src/sidetree/__tests__/test-utils.js
@@ -1,11 +1,11 @@
-const element = require('../../index');
+const element = require('../../../index');
 const {
   encodeJson,
   decodeJson,
   syncTransaction,
-} = require('./func');
-const { getDidDocumentModel, getCreatePayload } = require('./op');
-const { resolve } = require('./protocol');
+} = require('../func');
+const { getDidDocumentModel, getCreatePayload } = require('../op');
+const resolve = require('../resolve');
 
 const getTestSideTree = () => {
   const db = new element.adapters.database.ElementRXDBAdapter({
