+++ packages/element-lib/src/sidetree/batch/batchWrite.js
