+++ packages/element-lib/src/sidetree/op/index.js
@@ -1,4 +1,4 @@
-const { encodeJson, signEncodedPayload } = require('./func');
+const { encodeJson, signEncodedPayload } = require('../func');
 
 const getDidDocumentModel = (primaryPublicKey, recoveryPublicKey) => ({
   '@context': 'https://w3id.org/did/v1',
