+++ packages/element-lib/src/sidetree/sync/index.js
