+++ packages/element-lib/src/sidetree/utils/getNodeInfo.js
