+++ packages/element-lib/src/sidetree/utils/transactions.js
