+++ packages/element-lib/src/sidetree/utils/validation.spec.js
